atomic: Add git-core and buildah
This is useful for at least bootstrapping dev containers; and also in general I
think we should have lots of container tools in the base *workstation* host,
even if not everyone uses all of them.  Workstation is far from minimal today
anyways, and the high level goal is: give people lots of tools and means
to containerize.