Add comps-sync script, use it to trim package set
This script looks for packages not in comps at all, which
helps us lose things that were there only as dependencies.

Also, it now supports showing packages in the workstation environment
but not in the manifest.  (I'm not yet taking action on those though)

Some leaf packages do drop out. All of these I believe are right to stop
shipping by default with the possible exception of `media-player-info`...I'm not
sure if that was intentional or not. Anyways let's start this sync process.

```
-Installing 1344 packages:
+Installing 1335 packages:
-  bcache-tools-1.0.8-10.fc27.x86_64 (fedora-rawhide)
-  isomd5sum-1:1.2.2-1.fc28.x86_64 (fedora-rawhide)
-  kexec-tools-2.0.16-3.fc28.x86_64 (fedora-rawhide)
-  keybinder3-0.3.2-3.fc27.x86_64 (fedora-rawhide)
-  langtable-0.0.38-2.fc28.noarch (fedora-rawhide)
-  langtable-data-0.0.38-2.fc28.noarch (fedora-rawhide)
-  langtable-python3-0.0.38-2.fc28.noarch (fedora-rawhide)
-  lpsolve-5.5.2.0-16.fc27.x86_64 (fedora-rawhide)
-  media-player-info-23-1.fc28.noarch (fedora-rawhide)
```