comps-sync.py: Support multiple comps groups for a desktop
Enable fetching packages from multiple comps groups for each desktop.
This is currently unused but this removes the constraints around having
all base packages for a given desktop in a single comps group.

This could also be used to get packages from other groups such as the
input-methods group for example.