Drop some now-unneeded manual handling of arch-specific packages
This lets the recently-merged improved arch-specific package
support handle all the arch-specific packages which were
previously excluded in `comps-sync-exclude-list.yml` and listed
manually in `fedora-common-ostree.yaml`. As the diff shows,
the sync script now correctly includes the same packages for
the same arches in `fedora-common-ostree-pkgs.yaml`, plus a few
appropriate additions of things that should be there but had
been left out, on ppc64le and aarch64.

We also drop the `packages-armhfp` and `packages-ppc64` lists,
as we no longer build for either of those arches. This allows
us to move `ostree-grub2` into the non-arch-specific list, since
it's no longer left out on any arch we care about.

Signed-off-by: Adam Williamson <awilliam@redhat.com>