Replace references to qgnomeplatform with qgnomeplatform-qt5
There has been no package called just 'qgnomeplatform' for some
time. qgnomeplatform-qt5 used to provide 'qgnomeplatform', but
.fc39 builds onward do not, so now we have a .fc39 build of it,
this breaks.

I'm not sure all these places where we list qgnomeplatform
exactly make sense - we seem to be excluding it in some places
then explicitly pulling it in in others - but this is the
minimal change to fix the immediate problem.

Signed-off-by: Adam Williamson <awilliam@redhat.com>